requirements.txt
praw
python-dotenv
tqdm
google-generativeai
